swc_trace_macro

Attribute Macro swc_trace

#[swc_trace]

Utility proc macro to add #[tracing::instrument(level = "info", skip_all)] to all methods in an impl block.

This attribute macro is typically applied on an VisitMut impl block. If this is applied, all implemented methods will annotated with the instrument annotation from tracing.